02_authentic/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
PyJWT[crypto]==2.10.1
